Updated README.md with a warning (#303)
* Updated README.md with a warning

I spent a lot of time debugging the `onClick` for `MenuItem` and after trying again and again I found in the issue tracker a comment explaining that every single instance of `ContextMenu` and `ContextMenuTrigger` needs to have a unique `id` they cannot share one even if they are the same component or use the same menu. 

I think this a good change to make to the docs since it's very small and can help people prevent people from wasting time debugging. Thank you for your time.

* Update README.md